[03622] Extract git command execution and parsing into separate helpers
